Fix race condition when 2 proc created DB lock at the same time
Quite a difficult bug to fix although expected when trying to use such a
simple mechanism (lock file) across different processes.
It is not the most elegant thing ever, but it's simple, cross platform,
and it works.

+++ src/file_lock.cpp
@@ -9,13 +9,26 @@ FileLock::FileLock(const std::string& filename)
 
 FileLockGuard::FileLockGuard(std::unique_ptr<FileLock> lockFile)
     : m_lockFile(std::move(lockFile))
+    , m_removeFile(true)
 {
     m_lockFile->createFile();
 }
 
 FileLockGuard::~FileLockGuard()
 {
-    m_lockFile->removeFile();
+    if (m_removeFile) {
+        m_lockFile->removeFile();
+    }
+}
+
+FileLock* FileLockGuard::getLockFile() const
+{
+    return m_lockFile.get();
+}
+
+void FileLockGuard::removeFile(bool remove)
+{
+    m_removeFile = remove;
 }
 
 FileLockImpl::FileLockImpl(const std::string& filename)
@@ -25,10 +38,36 @@ FileLockImpl::FileLockImpl(const std::string& filename)
 
 void FileLockImpl::createFile()
 {
-    std::ofstream file(m_filename);
+    std::ofstream ofs(m_filename);
 }
 
 void FileLockImpl::removeFile()
 {
     std::remove(m_filename.c_str());
+}
+
+bool FileLockImpl::writeToFile(const std::string& from)
+{
+    std::ofstream ofs(m_filename, std::ios::app);
+    if (ofs.good()) {
+        ofs << from << '\n';
+        ofs.flush();
+        return true;
+    }
+    return false;
+}
+
+bool FileLockImpl::readFromFile(std::string& to)
+{
+    std::ifstream ifs(m_filename);
+    if (ifs.good()) {
+        ifs.seekg(0);
+        std::string line;
+        while (!ifs.eof()) {
+            std::getline(ifs, line);
+            to += line;
+        }
+        return true;
+    }
+    return false;
 }
